+++ doc/guide/iodoc.hpp
@@ -120,7 +120,8 @@ and debugging output for your mlpack program.
 @section simpleio Simple IO Example
 
 Through the mlpack::IO object, command-line parameters can be easily added
-with the PROGRAM_INFO, PARAM_INT, PARAM_DOUBLE, PARAM_STRING, and PARAM_FLAG
+with the BINDING_NAME, BINDING_SHORT_DESC, BINDING_LONG_DESC, BINDING_EXAMPLE,
+BINDING_SEE_ALSO, PARAM_INT, PARAM_DOUBLE, PARAM_STRING, and PARAM_FLAG
 macros.
 
 Here is a sample use of those macros, extracted from methods/pca/pca_main.cpp.
@@ -131,23 +132,28 @@ Here is a sample use of those macros, extracted from methods/pca/pca_main.cpp.
 #include <mlpack/core/util/io.hpp>
 #include <mlpack/core/util/mlpack_main.hpp>
 
-// Document program.
-PROGRAM_INFO("Principal Components Analysis",
-    // Short description.
+// Program Name.
+BINDING_NAME("Principal Components Analysis");
+
+// Short description.
+BINDING_SHORT_DESC(
     "An implementation of several strategies for principal components analysis "
     "(PCA), a common preprocessing step.  Given a dataset and a desired new "
     "dimensionality, this can reduce the dimensionality of the data using the "
-    "linear transformation determined by PCA.",
-    // Long description.
+    "linear transformation determined by PCA.");
+
+// Long description.
+BINDING_LONG_DESC(
     "This program performs principal components analysis on the given dataset "
     "using the exact, randomized, randomized block Krylov, or QUIC SVD method. "
     "It will transform the data onto its principal components, optionally "
     "performing dimensionality reduction by ignoring the principal components "
-    "with the smallest eigenvalues."
-    // "See also" section for generated documentation.
-    SEE_ALSO("Principal component analysis on Wikipedia",
-        "https://en.wikipedia.org/wiki/Principal_component_analysis"),
-    SEE_ALSO("mlpack::pca::PCA C++ class documentation",
+    "with the smallest eigenvalues.");
+
+// See also...
+BINDING_SEE_ALSO("Principal component analysis on Wikipedia",
+        "https://en.wikipedia.org/wiki/Principal_component_analysis");
+BINDING_SEE_ALSO("mlpack::pca::PCA C++ class documentation",
         "@doxygen/classmlpack_1_1pca_1_1PCA.html"));
 
 // Parameters for program.