+++ src/mlpack/methods/lmnn/lmnn_main.cpp
@@ -21,14 +21,18 @@
 
 #include <ensmallen.hpp>
 
-// Define parameters.
-PROGRAM_INFO("Large Margin Nearest Neighbors (LMNN)",
-    // Short description.
+// Program Name.
+BINDING_NAME("Large Margin Nearest Neighbors (LMNN)");
+
+// Short description.
+BINDING_SHORT_DESC(
     "An implementation of Large Margin Nearest Neighbors (LMNN), a distance "
     "learning technique.  Given a labeled dataset, this learns a transformation"
     " of the data that improves k-nearest-neighbor performance; this can be "
-    "useful as a preprocessing step.",
-    // Long description.
+    "useful as a preprocessing step.");
+
+// Long description.
+BINDING_LONG_DESC(
     "This program implements Large Margin Nearest Neighbors, a distance "
     "learning technique.  The method seeks to improve k-nearest-neighbor "
     "classification on a dataset.  The method employes the strategy of "
@@ -111,8 +115,10 @@ PROGRAM_INFO("Large Margin Nearest Neighbors (LMNN)",
     "literature on L-BFGS.  In addition, a normalized starting point can be "
     "used by specifying the " + PRINT_PARAM_STRING("normalize") + " parameter."
     "\n\n"
-    "By default, the AMSGrad optimizer is used."
-    "\n\n"
+    "By default, the AMSGrad optimizer is used.");
+
+// Example.
+BINDING_EXAMPLE(
     "Example - Let's say we want to learn distance on iris dataset with "
     "number of targets as 3 using BigBatch_SGD optimizer. A simple call for "
     "the same will look like: "
@@ -124,15 +130,17 @@ PROGRAM_INFO("Large Margin Nearest Neighbors (LMNN)",
     "with dataset having labels as last column can be made as: "
     "\n\n" +
     PRINT_CALL("mlpack_lmnn", "input", "letter_recognition", "k", 5,
-    "range", 10, "regularization", 0.4, "output", "output"),
-    SEE_ALSO("@nca", "#nca"),
-    SEE_ALSO("Large margin nearest neighbor on Wikipedia",
-        "https://en.wikipedia.org/wiki/Large_margin_nearest_neighbor"),
-    SEE_ALSO("Distance metric learning for large margin nearest neighbor "
+    "range", 10, "regularization", 0.4, "output", "output"));
+
+// See also...
+BINDING_SEE_ALSO("@nca", "#nca");
+BINDING_SEE_ALSO("Large margin nearest neighbor on Wikipedia",
+        "https://en.wikipedia.org/wiki/Large_margin_nearest_neighbor");
+BINDING_SEE_ALSO("Distance metric learning for large margin nearest neighbor "
         "classification (pdf)", "http://papers.nips.cc/paper/2795-distance-"
-        "metric-learning-for-large-margin-nearest-neighbor-classification.pdf"),
-    SEE_ALSO("mlpack::lmnn::LMNN C++ class documentation",
-        "@doxygen/classmlpack_1_1lmnn_1_1LMNN.html"));
+        "metric-learning-for-large-margin-nearest-neighbor-classification.pdf");
+BINDING_SEE_ALSO("mlpack::lmnn::LMNN C++ class documentation",
+        "@doxygen/classmlpack_1_1lmnn_1_1LMNN.html");
 
 PARAM_MATRIX_IN_REQ("input", "Input dataset to run LMNN on.", "i");
 PARAM_MATRIX_IN("distance", "Initial distance matrix to be used as "