+++ HISTORY.md
@@ -1,5 +1,12 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Force CMake to show error when it didn't find Python/modules (#2568).
+
+  * Refactor `ProgramInfo()` to separate out all the different
+    information (#2558).
+
+  * Added Soft Actor-Critic to RL methods (#2487).
+
   * Added Categorical DQN to q_networks (#2454).
 
   * Added N-step DQN to q_networks (#2461).
@@ -17,6 +24,10 @@
     Julia, and Go.  Also called "Bayesian Ridge", this is equivalent to a
     version of linear regression where the regularization parameter is
     automatically tuned (#2030).
+
+  * Fix incremental training of logistic regression models (#2560).
+
+  * Change default configuration of `BUILD_PYTHON_BINDINGS` to `OFF` (#2575).
 
 ### mlpack 3.3.2
 ###### 2020-06-18